JUS-25: refactor: Extract types
